Update fan.py
it appears the speeds as reported via telnet have changed.  the decimal is now 0 for all speeds.  Low = 25.00, Medium = 50.00, Medium High = 75.00, and High = 100.00.  I have modified the file in my own home assistant and tested and it correctly sets the speed and also reports the correct speed if changed externally outside of HA.

bug reports:
https://github.com/upsert/lutron-caseta-pro/issues/72
https://github.com/home-assistant/core/issues/44377